Create the necessary infrastructure for users CRUD.
Using the examples from this lesson, secure your blog application such that you have to log in to create a post.

Test this functionality, ensure that you get redirected to the login page if you try to create a post without being logged in.

When a post is created, assign the logged in user as the owner of that post.

+++ src/main/java/com/codeup/repositories/PostsRepository.java
@@ -1,20 +1,22 @@
 package com.codeup.repositories;
 
-import com.codeup.models.Post;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
 
 
 /**
  * Created by vanessamnoble on 2/9/17.
  */
 
-//insert -> table name, values, columns
-//select -> table, columns, values
-//update -> table, columns, values
-//delete -> table, values
+
+
+import com.codeup.models.Post;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+// insert -> name, values, columns
+// select -> table, columns, values
+// update -> table, columns, values
+// delete -> table, values
 
 @Repository
-public interface PostsRepository extends CrudRepository<Post, Long> {
-
+public interface PostsRepository extends CrudRepository<Post, Integer> {
 }
